tutorial/tutorial_to_replicate/requirements.txt
google-cloud-bigquery[pandas]>=3.0,<4.0
pandas
pandas_gbq
fsspec
gcsfs
